requirements.txt
asgiref==3.6.0
Django==4.1.7
dj-database-url==0.5.0
django-allauth==0.42.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.1
drf-spectacular==0.26.1
#Pillow==8.1.0
urllib3==1.26.4
gunicorn==20.1.0
PyJWT==2.6.0
pytz==2022.7.1
requests==2.25.1
sqlparse==0.4.3
whitenoise==6.3.0
